PIN assignment bugs
Fixed a bug which used the preprocessor defined pins instead of the class instance variables.   The problem is that this was determined at compile time, which was before the user code was built.  So there was not way to configure the pins to anything other than the preprocessor defaults.

Note, if you rebuilt the binary it would work, with the correct includes.  But that was not as advertised.